refactor: localStorage 관리 함수 한 곳에 모아놔라
