+++ src/app/localStorage/coin.js
@@ -1,5 +1,5 @@
 import { COIN, STORAGE_KEY } from '../asset/constants/index.js';
-import { getJsonItem, setJsonItem } from './index.js';
+import { getJsonItem, setJsonItem } from './common.js';
 import { calcCoinAmount } from '../asset/util/index.js';
 
 const defaultCoinVal = COIN.map(() => 0);
