+++ src/app/localStorage/inventory.js
@@ -1,7 +1,9 @@
 import { STORAGE_KEY } from '../asset/constants/index.js';
-import { getJsonItem, setJsonItem, appendToJsonItem } from './index.js';
+import { getJsonItem, setJsonItem, appendToJsonItem } from './common.js';
 
 export const getProducts = () => getJsonItem(STORAGE_KEY.inventory);
+
+export const setProducts = (products) => setJsonItem(STORAGE_KEY.inventory, products);
 
 export const addProduct = (productName, productPrice, productQuantity) => {
     appendToJsonItem(STORAGE_KEY.inventory, {
@@ -10,8 +12,6 @@ export const addProduct = (productName, productPrice, productQuantity) => {
         productQuantity,
     });
 };
-
-export const setProducts = (products) => setJsonItem(STORAGE_KEY.inventory, products);
 
 export const isUniqueProductName = (candProductName) =>
     getProducts().findIndex(({ productName }) => productName === candProductName) === -1;
